Chunk mesh error fix + address compiler warnings

+++ src/game/world_gen.cpp
@@ -7,12 +7,12 @@ namespace Generator {
     }
 
     int GetXZHeight(int x, int z) {
-        float val = Noise::simplex2(x * 0.01, z * 0.01, 2, 2, 2);
-        return 4 + (val * 16);
+        float val = Noise::simplex2(x * 0.01f, z * 0.01f, 2, 2, 2);
+        return 4 + (int)(val * 16);
     }
 
     bool IsTree(int x, int z) {
-        float val = Noise::simplex2(x, z, 8, 1, 1);
+        float val = Noise::simplex2((float)x, (float)z, 8, 1, 1);
         if (val >= 0.998) {
             return true;
         }
